+++ src/Services/Algolia.php
@@ -42,59 +42,6 @@ class Algolia
 	}
 
 	/**
-	 * @deprecated use uploadValues instead
-	 * @param array<string, array<string, array<string>>> $indexes
-	 * @param array<string> $mutations
-	 * @throws \Algolia\AlgoliaSearch\Exceptions\MissingObjectId|\StORM\Exception\NotFoundException
-	 */
-	public function uploadProducts(array $indexes, array $mutations): void
-	{
-		if (!$this->client) {
-			return;
-		}
-
-		$category = ($categorySetting = $this->settingRepository->one(['name' => 'algoliaCategory'])) ? $this->categoryRepository->one($categorySetting->value) : null;
-
-		$algoliaResults = [];
-
-		foreach ($indexes as $indexName => $indexData) {
-			$properties = $indexData['properties'];
-			$mutationalProperties = $indexData['mutationalProperties'];
-
-			foreach ($mutations as $mutation) {
-				$records = $this->productRepository->many();
-
-				if ($category) {
-					$records->join(['productXCategory' => 'eshop_product_nxn_eshop_category'], 'this.uuid = productXCategory.fk_product')
-						->join(['category' => 'eshop_category'], 'productXCategory.fk_category = category.uuid')
-						->where('category.path LIKE :s', ['s' => "$category->path%"]);
-				}
-
-				/** @var \Eshop\DB\Product $record */
-				foreach ($records as $record) {
-					$algoliaResults[$record->getPK()]['objectID' ] = $record->getPK();
-
-					foreach ($properties as $property) {
-						$algoliaResults[$record->getPK()][$property] = ($property === 'page_url' ? $this->baseUrl : '') . $record->$property;
-						$algoliaResults[$record->getPK()]['not_' . $property] = (bool) !$record->$property;
-					}
-
-					foreach ($mutationalProperties as $property) {
-						$prop = $property . '_' . $mutation;
-						$algoliaResults[$record->getPK()][$property . '_' . $mutation] = ($property === 'page_url' ? $this->baseUrl : '') . $record->$prop;
-						$algoliaResults[$record->getPK()]['not_' . $property . '_' . $mutation] = (bool) !$record->$prop;
-					}
-				}
-			}
-
-			$index = $this->client->initIndex($indexName);
-			$index->saveObjects($algoliaResults, [
-				'objectIDKey' => 'objectID',
-			]);
-		}
-	}
-
-	/**
 	 * @deprecated use search instead
 	 * @param string $name
 	 * @param string $index